Fixed build script

+++ build/build.ps1
@@ -1,2 +1,2 @@
-Remove-Item ..\output -Recurse
+if(Test-Path ..\output) { Remove-Item ..\output -Recurse }
 ..\nuget\nuget.exe pack ..\src\TestTemplate.nuspec -OutputDirectory ..\output
